+++ backtest/strategy.go
@@ -1,7 +1,7 @@
 package backtest
 
 import (
-	"github.com/tsuz/go-pine"
+	"github.com/tsuz/go-pine/pine"
 )
 
 type Strategy interface {
