+++ pine/series_change_test.go
