+++ pine/series_ema.go
