+++ pine/series_ema_test.go
