+++ pine/series_macd_test.go
