+++ pine/series_pow.go
