+++ pine/series_pow_test.go
