+++ pine/series_roc_test.go
