+++ pine/series_sma.go
