+++ pine/series_sma_test.go
