+++ pine/series_stdev.go
