+++ pine/series_stdev_test.go
