+++ pine/series_sum_test.go
