+++ pine/series_value_when_test.go
