+++ pine/series_variance.go
