+++ pine/series_variance_test.go
