+++ pine/value_series_test.go
